@@ -17,8 +17,8 @@
 		}
 	],
 	"require": {
-		"php": "^7.4.0 || ^8.0.0",
-		"psr/log": "^1.1.0",
+		"php": "^8.3.0",
+		"psr/log": "^3.0.0",
 		"ext-curl": "*",
 		"ext-json": "*"
 	},
